[FLINK-30827][sql-client] Improve the SQL Client code structure
