+++ flink-table/flink-sql-client/src/main/java/org/apache/flink/table/client/gateway/result/ChangelogResult.java
@@ -16,15 +16,16 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.client.gateway.local.result;
+package org.apache.flink.table.client.gateway.result;
 
-/**
- * A result of a dynamic table program.
- *
- * <p>Note: Make sure to call close() after the result is not needed anymore.
- */
-public interface DynamicResult {
+import org.apache.flink.table.client.gateway.TypedResult;
+import org.apache.flink.table.data.RowData;
 
-    /** Closes the retrieval and all involved threads. */
-    void close();
+import java.util.List;
+
+/** A result that is represented as a changelog consisting of insert and delete records. */
+public interface ChangelogResult extends DynamicResult {
+
+    /** Retrieves the available result records. */
+    TypedResult<List<RowData>> retrieveChanges();
 }
