+++ flink-table/flink-sql-client/src/main/java/org/apache/flink/table/client/gateway/result/DynamicResult.java
@@ -16,22 +16,15 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.client.gateway.local.result;
+package org.apache.flink.table.client.gateway.result;
 
-import org.apache.flink.table.client.gateway.TypedResult;
-import org.apache.flink.table.data.RowData;
+/**
+ * A result of a dynamic table program.
+ *
+ * <p>Note: Make sure to call close() after the result is not needed anymore.
+ */
+public interface DynamicResult {
 
-import java.util.List;
-
-/** A result that is materialized and can be viewed by navigating through a snapshot. */
-public interface MaterializedResult extends DynamicResult {
-
-    /**
-     * Takes a snapshot of the current table and returns the number of pages for navigating through
-     * the snapshot.
-     */
-    TypedResult<Integer> snapshot(int pageSize);
-
-    /** Retrieves a page of a snapshotted result. */
-    List<RowData> retrievePage(int page);
+    /** Closes the retrieval and all involved threads. */
+    void close();
 }
