+++ flink-table/flink-sql-client/src/main/java/org/apache/flink/table/client/gateway/result/MaterializedCollectStreamResult.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.client.gateway.local.result;
+package org.apache.flink.table.client.gateway.result;
 
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.table.client.gateway.StatementResult;
