+++ flink-table/flink-sql-client/src/test/java/org/apache/flink/table/client/gateway/result/BaseMaterializedResultTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.client.gateway.local.result;
+package org.apache.flink.table.client.gateway.result;
 
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.data.binary.BinaryRowData;
